liBlog/blogs/search_indexes.py
import datetime
from haystack import indexes
from .models import Post


class PostIndex(indexes.SearchIndex, indexes.Indexable):
    text = indexes.CharField(document=True, use_template=True)

    publish_time = indexes.DateTimeField(model_attr='publish_time')

    def get_model(self):
        return Post

    def index_queryset(self, using=None):
        """Used when the entire index for model is updated."""
        return self.get_model().objects.filter(
            publish_time__lte=datetime.datetime.now())
