requirements.txt
amqp==1.4.9
anyjson==0.3.3
argon2-cffi==16.3.0
awesome-slugify==1.6.5
billiard==3.3.0.23
celery==3.1.25
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
Cython==0.28.2
defusedxml==0.5.0
Django==1.11.5
django-allauth==0.36.0
django-appconf==1.0.2
django-baker==0.21
django-ckeditor==5.4.0
django-comments-xtd==2.1.0
django-compressor==2.2
django-contrib-comments==1.8.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-haystack==2.8.1
django-js-asset==1.0.0
django-markdown2==0.3.1
django-mdeditor==0.1.3
django-model-utils==3.0.0
django-redis==4.8.0
django-threadedcomments==1.2
djangorestframework==3.8.2
docutils==0.14
geoip2==2.8.0
gevent==1.3.1
greenlet==0.4.13
gunicorn==19.8.1
idna==2.6
jieba==0.39
kombu==3.0.37
markdown2==2.3.5
maxminddb==1.3.0
mistune==0.8.3
mistune-contrib==0.1
oauthlib==2.0.7
olefile==0.45.1
Pillow==4.2.1
pycparser==2.18
Pygments==2.2.0
PyMySQL==0.8.0
python3-openid==3.1.0
pytz==2017.2
rcssmin==1.0.6
redis==2.10.6
regex==2018.2.21
requests==2.18.4
requests-oauthlib==0.8.0
rjsmin==1.0.12
six==1.11.0
sqlparse==0.2.4
Unidecode==0.4.21
urllib3==1.22
whitenoise==3.3.1
Whoosh==2.7.4